# contracts/openapi/pms-api.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PMS Backend API",
    "version": "1.0.0",
    "description": "Contrat initial publié par pms-backend pour le BFF pms-frontend."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "security": [
    {
      "keycloakBearer": []
    }
  ],
  "paths": {
...
    "/products": {
      "post": {
        "operationId": "createInvestmentProduct",
        "parameters": [
          {
            "$ref": "#/components/parameters/CorrelationId"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvestmentProduct"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Produit créé en brouillon",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestmentProduct"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/products/{productId}": {
      "get": {
        "operationId": "getInvestmentProduct",
        "parameters": [
          {
            "$ref": "#/components/parameters/CorrelationId"
          },
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Produit d'investissement",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestmentProduct"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/products/{productId}/{action}": {
      "post": {
        "operationId": "transitionInvestmentProduct",
        "parameters": [
          {
            "$ref": "#/components/parameters/CorrelationId"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "action",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "validate",
                "publish",
                "suspend",
                "resume",
                "close"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductTransitionCommand"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transition appliquée et auditée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestmentProduct"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/products/{productId}/terms/simulate": {
      "post": {
        "operationId": "simulateProductTerms",
        "parameters": [
          {
            "$ref": "#/components/parameters/CorrelationId"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductTermsDraft"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Simulation sans mutation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductTermsSimulation"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/products/{productId}/terms": {
      "post": {
        "operationId": "createProductTermsDraft",
        "parameters": [
          {
            "$ref": "#/components/parameters/CorrelationId"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductTermsDraft"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Version contractuelle en brouillon",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductTerms"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/products/{productId}/terms/{termsVersionId}/publish": {
      "post": {
        "operationId": "publishProductTerms",
        "parameters": [
          {
            "$ref": "#/components/parameters/CorrelationId"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "termsVersionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishProductTerms"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Version contractuelle publiée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductTerms"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/products/compliance-references": {
      "post": {
        "operationId": "createComplianceReference",
        "parameters": [
          {
            "$ref": "#/components/parameters/CorrelationId"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateComplianceReference"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Référence de conformité versionnée créée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplianceReference"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/products/{productId}/references": {
      "get": {
        "operationId": "listProductReferences",
        "parameters": [
          {
            "$ref": "#/components/parameters/CorrelationId"
          },
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Références produit ordonnées par autorité",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductReference"
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      },
      "post": {
        "operationId": "associateProductReference",
        "parameters": [
          {
            "$ref": "#/components/parameters/CorrelationId"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssociateProductReference"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Référence associée"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/products/{productId}/compliance-arbitrations": {
      "post": {
        "operationId": "arbitrateProductCompliance",
        "parameters": [
          {
            "$ref": "#/components/parameters/CorrelationId"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComplianceArbitration"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Décision d'arbitrage append-only"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
    "/reporting/regulatory-reports/generate": {
      "post": {
        "operationId": "generateRegulatoryReport",
        "parameters": [
          {
            "$ref": "#/components/parameters/CorrelationId"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegulatoryReportGeneration"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Rapport réglementaire généré",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratedRegulatoryReport"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/reporting/regulatory-reports/{regulatoryReportId}/publish": {
      "post": {
        "operationId": "publishRegulatoryReport",
        "parameters": [
          {
            "$ref": "#/components/parameters/CorrelationId"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "name": "regulatoryReportId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegulatoryReportPublication"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Rapport réglementaire publié",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegulatoryReportTransition"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "keycloakBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "CorrelationId": {
        "name": "X-Correlation-ID",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 16,
          "maxLength": 128
        }
      }
    },
    "responses": {
      "Problem": {
        "description": "Erreur normalisée RFC 9457",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      }
    },
    "schemas": {
...
      "CreateInvestmentProduct": {
        "type": "object",
        "required": [
          "code",
          "name",
          "investorNisba",
          "bankNisba"
        ],
        "properties": {
          "code": {
            "type": "string",
            "pattern": "^[A-Z0-9_-]{2,32}$"
          },
          "name": {
            "type": "string",
            "minLength": 3
          },
          "investorNisba": {
            "type": "string",
            "pattern": "^(?:0|[1-9][0-9]{0,2})(?:\\.[0-9]{1,6})?$"
          },
          "bankNisba": {
            "type": "string",
            "pattern": "^(?:0|[1-9][0-9]{0,2})(?:\\.[0-9]{1,6})?$"
          },
          "shariaReference": {
            "type": "string",
            "minLength": 1
          }
        },
        "additionalProperties": false
      },
      "ProductTransitionCommand": {
        "type": "object",
        "required": [
          "justification"
        ],
        "properties": {
          "justification": {
            "type": "string",
            "minLength": 10,
            "maxLength": 1000
          }
        },
        "additionalProperties": false
      },
      "InvestmentProduct": {
        "type": "object",
        "required": [
          "productId",
          "code",
          "name",
          "investorNisba",
          "bankNisba",
          "status"
        ],
        "properties": {
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "investorNisba": {
            "type": "string"
          },
          "bankNisba": {
            "type": "string"
          },
          "shariaReference": {
            "type": "string"
          },
          "validatedBy": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "VALIDATED",
              "PUBLISHED",
              "SUSPENDED",
              "CLOSED"
            ]
          }
        },
        "additionalProperties": false
      },
      "ProductTermsDraft": {
        "type": "object",
        "required": [
          "effectiveFrom",
          "investorNisba",
          "bankNisba"
        ],
        "properties": {
          "effectiveFrom": {
            "type": "string",
            "format": "date"
          },
          "effectiveTo": {
            "type": "string",
            "format": "date"
          },
          "investorNisba": {
            "type": "string"
          },
          "bankNisba": {
            "type": "string"
          },
          "indicativeTargetRate": {
            "type": "string",
            "description": "Taux cible indicatif et non garanti"
          }
        },
        "additionalProperties": false
      },
      "ProductTermsSimulation": {
        "type": "object",
        "required": [
          "simulationChecksumSha256",
          "valid",
          "notice"
        ],
        "properties": {
          "simulationChecksumSha256": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$"
          },
          "valid": {
            "type": "boolean",
            "const": true
          },
          "notice": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "PublishProductTerms": {
        "type": "object",
        "required": [
          "businessDate",
          "simulationChecksumSha256",
          "justification"
        ],
        "properties": {
          "businessDate": {
            "type": "string",
            "format": "date"
          },
          "simulationChecksumSha256": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$"
          },
          "retroactiveApprovalId": {
            "type": "string",
            "minLength": 16,
            "maxLength": 128
          },
          "justification": {
            "type": "string",
            "minLength": 10,
            "maxLength": 1000
          }
        },
        "additionalProperties": false
      },
      "ProductTerms": {
        "type": "object",
        "required": [
          "termsVersionId",
          "productId",
          "version",
          "effectiveFrom",
          "investorNisba",
          "bankNisba",
          "status",
          "createdBy"
        ],
        "properties": {
          "termsVersionId": {
            "type": "string",
            "format": "uuid"
          },
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "version": {
            "type": "integer",
            "minimum": 1
          },
          "effectiveFrom": {
            "type": "string",
            "format": "date"
          },
          "effectiveTo": {
            "type": "string",
            "format": "date"
          },
          "investorNisba": {
            "type": "string"
          },
          "bankNisba": {
            "type": "string"
          },
          "indicativeTargetRate": {
            "type": "string",
            "description": "Taux cible indicatif et non garanti"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "PUBLISHED",
              "RETIRED"
            ]
          },
          "simulationChecksumSha256": {
            "type": "string"
          },
          "retroactiveApprovalId": {
            "type": "string"
          },
          "createdBy": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "CreateComplianceReference": {
        "type": "object",
        "required": [
          "source",
          "referenceCode",
          "version",
          "title",
          "effectiveFrom"
        ],
        "properties": {
          "source": {
            "type": "string",
            "enum": [
              "BA",
              "SHARIA_COMMITTEE",
              "AAOIFI",
              "IFSB"
            ]
          },
          "referenceCode": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "effectiveFrom": {
            "type": "string",
            "format": "date"
          },
          "effectiveTo": {
            "type": "string",
            "format": "date"
          }
        },
        "additionalProperties": false
      },
      "ComplianceReference": {
        "type": "object",
        "required": [
          "referenceId",
          "source",
          "referenceCode",
          "version",
          "title",
          "effectiveFrom",
          "createdBy"
        ],
        "properties": {
          "referenceId": {
            "type": "string",
            "format": "uuid"
          },
          "source": {
            "type": "string",
            "enum": [
              "BA",
              "SHARIA_COMMITTEE",
              "AAOIFI",
              "IFSB"
            ]
          },
          "referenceCode": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "effectiveFrom": {
            "type": "string",
            "format": "date"
          },
          "effectiveTo": {
            "type": "string",
            "format": "date"
          },
          "createdBy": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "ProductReference": {
        "type": "object",
        "required": [
          "referenceId",
          "source",
          "referenceCode",
          "version",
          "title",
          "effectiveFrom",
          "createdBy",
          "kind",
          "associatedAt"
        ],
        "properties": {
          "referenceId": {
            "type": "string",
            "format": "uuid"
          },
          "source": {
            "type": "string",
            "enum": [
              "BA",
              "SHARIA_COMMITTEE",
              "AAOIFI",
              "IFSB"
            ]
          },
          "referenceCode": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "effectiveFrom": {
            "type": "string",
            "format": "date"
          },
          "effectiveTo": {
            "type": "string",
            "format": "date"
          },
          "createdBy": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "enum": [
              "CONTRACTUAL_DOCUMENT",
              "REGULATORY_DOCUMENT",
              "SHARIA_DOCUMENT",
              "ACCOUNTING_SCHEMA"
            ]
          },
          "associatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "AssociateProductReference": {
        "type": "object",
        "required": [
          "referenceId",
          "kind"
        ],
        "properties": {
          "referenceId": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "type": "string",
            "enum": [
              "CONTRACTUAL_DOCUMENT",
              "REGULATORY_DOCUMENT",
              "SHARIA_DOCUMENT",
              "ACCOUNTING_SCHEMA"
            ]
          }
        },
        "additionalProperties": false
      },
...
      "ComplianceArbitration": {
        "type": "object",
        "required": [
          "selectedReferenceId",
          "rejectedReferenceId",
          "rationale"
        ],
        "properties": {
          "selectedReferenceId": {
            "type": "string",
            "format": "uuid"
          },
          "rejectedReferenceId": {
            "type": "string",
            "format": "uuid"
          },
          "rationale": {
            "type": "string",
            "minLength": 20
          }
        },
        "additionalProperties": false
      },
...
      "RegulatoryReportGeneration": {
        "type": "object",
        "required": [
          "reportType",
          "period"
        ],
        "properties": {
          "reportType": {
            "type": "string",
            "pattern": "^[A-Z][A-Z0-9_-]{1,63}$"
          },
          "period": {
            "type": "string",
            "pattern": "^[0-9]{4}-(0[1-9]|1[0-2])$"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}